+++ operator/pkg/resources/proxy/statefulset.go
@@ -134,8 +134,8 @@ func proxyPodSpec(ais *aisv1.AIStore) *corev1.PodSpec {
 	if ais.Spec.PriorityClassName != nil {
 		spec.PriorityClassName = *ais.Spec.PriorityClassName
 	}
-	if ais.Spec.LogSidecarImage != nil {
-		spec.Containers = append(spec.Containers, cmn.NewLogSidecar(*ais.Spec.LogSidecarImage, aisapc.Proxy, ais.Spec.LogSidecarResources))
+	if ais.GetLogSidecarImage() != "" {
+		spec.Containers = append(spec.Containers, cmn.NewLogSidecar(ais, aisapc.Proxy))
 	}
 	return spec
 }